ADD: Testing for multi image requests with preprocessing images
