+++ spec/indico_spec.rb
@@ -160,19 +160,25 @@ describe Indico do
   it "should respond with all image apis called" do
     test_image = Array.new(48){Array.new(48){Array.new(3){rand(100)/100.0}}}
     expected_keys = Set.new(IMAGE_APIS)
-    response = Indico.predict_image(test_image, IMAGE_APIS)
+    silent_warnings do
+      response = Indico.predict_image(test_image, IMAGE_APIS)
 
-    expect(response.class).to eql(Hash)
-    expect(Set.new(response.keys)).to eql(expected_keys)
+      expect(response.class).to eql(Hash)
+      expect(Set.new(response.keys)).to eql(expected_keys)
+    end
   end
 
-  it "should respond with all text apis called in batch" do
-    expected_keys = Set.new(TEXT_APIS)
-    response = Indico.batch_predict_text(["Worst movie ever."], TEXT_APIS)
+  it "should respond with all image apis called on int array" do
+    test_image = Array.new(48){Array.new(48){Array.new(3){rand(100)}}}
+    expected_keys = Set.new(IMAGE_APIS)
+    silent_warnings do
+      response = Indico.predict_image(test_image, IMAGE_APIS)
 
-    expect(response.class).to eql(Hash)
-    expect(Set.new(response.keys)).to eql(expected_keys)
+      expect(response.class).to eql(Hash)
+      expect(Set.new(response.keys)).to eql(expected_keys)
+    end
   end
+
 
   # Uncomment when frontend updated to accept image urls
   # it 'should accept image urls' do
